+++ Core/Models/Economy/Banking/IBankATM.cs
@@ -8,7 +8,7 @@ namespace IgiCore.Core.Models.Economy.Banking
         DateTime Created { get; set; }
         DateTime? Deleted { get; set; }
         string Name { get; set; }
-        uint Hash { get; set; }
+        long Hash { get; set; }
 
         float PosX { get; set; }
         float PosY { get; set; }
